A big JSON update.
Now the serialization functions trim BigDecimal ending 
zeros and finaly make JSON.encode's output uniform. It
matters because persistent JSON strings must be
uniformely hashable to be referenced by distributed
and eventually disconnected network applications.

Also, I could not resist the convenience of adding support 
for Rhino native types in the JSON serialization support.
Anyway, who would want JSON without JavaScript instead of
Java? The performance impact is insignificant and. 

Once more, scripting practicality beats the shit of static 
purity and offers a place to put a very ugly patch on 
yet another broken Java spec.

To Note: BigDecimal.toString() is broken and will happilly 
print out "1.000000" if that is what was typed in. 

  "1.0000" != "1"
  "1.0001" != "1.000100"

Which means that Java's BigDecimal strings are not 
comparable to the text representation of integers
or various level of precision of the same decimal number!

Does it matter?

If you need to compare two JSON object by the equality of
the digest of their uniform textual representation ...

... YES!